requirements.txt
aiohttp~=3.8.1
discord-components~=2.1.2
Pillow~=8.4.0
dislash.py~=1.4.9
discord-py-slash-command==3.0.3
aiocron
